requirements.txt
anyio==3.7.0 
exceptiongroup==1.1.1 
fastapi==0.97.0 
idna==3.4 
pydantic==1.10.9 
sniffio==1.3.0 
starlette==0.27.0 
typing-extensions==4.6.3
greenlet==2.0.2
sqlalchemy==2.0.16
aiohttp==3.8.4 
aiosignal==1.3.1 
async-timeout==4.0.2 
attrs==23.1.0 
charset-normalizer==3.1.0 
frozenlist==1.3.3 
multidict==6.0.4 
yarl==1.9.2
asyncpg==0.27.0
iniconfig==2.0.0 
packaging==23.1 
pluggy==1.0.0 
pytest==7.3.2 
pytest-asyncio==0.21.0 
tomli==2.0.1
python-dotenv==1.0.0
apscheduler==3.10.1 
pytz==2023.3 
six==1.16.0 
tzlocal==5.0.1
click==8.1.3 
h11==0.14.0 
uvicorn==0.22.0
